rag-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
qdrant-client>=1.16.2
google-generativeai>=0.3.1
numpy>=1.24.3
pydantic>=2.5.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
PyPDF2>=3.0.0